@@ -1,10 +1,9 @@
-class CreateItems < ActiveRecord::Migration
+class CreateItems < ActiveRecord::Migration[5.0]
   def change
     create_table :items do |t|
       t.string :name
       t.references :list
-
-      t.timestamps null: false
+      t.timestamps
     end
   end
 end
