@@ -5,8 +5,8 @@ class ApiController < ApplicationController
   private
 
   def authenticated?
-    authenticate_or_request_with_http_basic { |username, password|
-      User.where( username: username ).present?
+    authenticate_or_request_with_http_basic {|username, password|
+      User.where(email: username).present? ? (User.where(email: username).first).valid_password?(password) : false
     }
   end
 
