@@ -1,4 +1,6 @@
 class List < ActiveRecord::Base
+  has_many :items, dependent: :destroy
   belongs_to :user
-  has_many :items
+
+  validates :name, length:{ minimum: 2, maximum: 100 }, presence: true
 end
