@@ -1,21 +1,11 @@
 class User < ActiveRecord::Base
+  before_save { self.email = email.downcase if email.present? }
+
+  validates :email, length: { minimum: 4, maximum: 100 }, presence: true
+
+  has_many :lists, dependent: :destroy
   # Include default devise modules. Others available are:
   # :confirmable, :lockable, :timeoutable and :omniauthable
   devise :database_authenticatable, :registerable,
          :recoverable, :rememberable, :trackable, :validatable
-
-
-  before_save { self.email = email.downcase if email.present? }
-
-  # validates :name, length: { minimum: 1, maximum: 100 }, presence: true
-  # validates :password, presence: true, length: { minimum: 6 }, unless: :password_digest
-  # validates :password, length: { minimum: 6 }, allow_blank: true
-  # validates :email,
-  #           presence: true,
-  #           uniqueness: { case_sensitive: false },
-  #           length: { minimum: 3, maximum: 254 }
-
-  # has_secure_password
-
-
 end
